Safari-611.1.10/LayoutTests/accessibility/gtk/text-for-range-table-cells-expected.txt

a table cell		
Block span in a link in a table cell
Inline span in a link in a table cell	
Block span in a table cell
Inline span in a table cell
This tests the ability to get element text for a specified range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cell1.stringForRange(0, -1) is 'a table cell'
PASS cell2.stringForRange(0, -1) is ''
PASS cell3.stringForRange(0, -1) is 'Block span in a link in a table cell\nInline span in a link in a table cell'
PASS cell4.stringForRange(0, -1) is 'Block span in a table cell\nInline span in a table cell'
PASS successfullyParsed is true

TEST COMPLETE
